refactor(vsock-host): isolate malformed command tests
